Use codegen to generate file.
