--- jack-sys/functions-gen/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-edition = "2021"
-name = "functions-gen"
-version = "0.1.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bitflags = "1"
